Return a championship from the ErgastClient class
ErgastClient exposed methods to get races, drivers, teams and qualy/race results.
There is no point in exposing all those, the clients of the ErgastClient class are only interested in getting all of those to build a championship object. Moving that responsibility to the ErgastClient class greatly simplifies the code and the tests.